renaming param.[ch] to plist.[ch]
since it seems more apt now that the list is identified by struct plist instead of struct param.
